requirements.txt
django==1.11.3
django-environ==0.4.4
gunicorn==19.7.1
mysqlclient==1.3.12
coverage==4.4.2
Pillow==4.2.1
django-allauth==0.34.0
django_crispy_forms==1.7.0

